+++ src/sender/mod.rs
@@ -1,2 +1,3 @@
+pub mod default;
 #[cfg(feature = "server")]
 pub mod server;
