+++ src/sender/server.rs
@@ -1,5 +1,4 @@
-use std::{io, net::SocketAddr, sync::Arc};
-
+use crate::utils::config::{FileMeta, BEAM_CLIENT, CONFIG};
 use axum::{
     extract::{Path, Request, State},
     http::{HeaderMap, StatusCode},
@@ -12,10 +11,10 @@ use axum_extra::{
 };
 use beam_lib::AppId;
 use futures_util::TryStreamExt as _;
+use std::{io, net::SocketAddr, sync::Arc};
 use tokio::net::TcpListener;
 use tokio_util::io::StreamReader;
-
-use crate::{FileMeta, BEAM_CLIENT, CONFIG};
+use tracing::error;
 
 pub async fn serve(addr: &SocketAddr, api_key: &str) -> anyhow::Result<()> {
     let app = Router::new()
@@ -39,12 +38,15 @@ async fn send_file(
     if auth.password() != api_key.as_ref() {
         return Err(StatusCode::UNAUTHORIZED);
     }
-    let mut parts = CONFIG.beam_id.as_ref().splitn(3, '.');
-    let app = parts.next().unwrap();
-    let _this_proxy = parts.next().unwrap();
-    let broker = parts.next().unwrap();
-
-    let to = AppId::new_unchecked(format!("{app}.{other_proxy_name}.{broker}"));
+    let to = AppId::new_unchecked(format!(
+        "{other_proxy_name}.{}",
+        CONFIG
+            .beam_id
+            .as_ref()
+            .rsplit('.')
+            .next()
+            .expect("AppId invalid"),
+    ));
     let mut conn = BEAM_CLIENT
         .create_socket_with_metadata(
             &to,
@@ -52,7 +54,7 @@ async fn send_file(
                 meta: headers.get("metadata").and_then(|v| {
                     serde_json::from_slice(v.as_bytes())
                         .map_err(|e| {
-                            eprintln!("Failed to deserialize metadata: {e}. Skipping metadata")
+                            error!("Failed to deserialize metadata: {e}. Skipping metadata")
                         })
                         .ok()
                 }),
@@ -63,7 +65,7 @@ async fn send_file(
         )
         .await
         .map_err(|e| {
-            eprintln!("Failed to tunnel request: {e}");
+            error!("Failed to tunnel request: {e}");
             StatusCode::INTERNAL_SERVER_ERROR
         })?;
     tokio::spawn(async move {
@@ -74,7 +76,7 @@ async fn send_file(
         );
         if let Err(e) = tokio::io::copy(&mut reader, &mut conn).await {
             // TODO: Some of these are normal find out which
-            eprintln!("Error sending file: {e}")
+            error!("Error sending file: {e}")
         }
     });
     Ok(())
